# backend-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Login with Google access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with Google access token",
                "parameters": [
                    {
                        "description": "Provider access token payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ProviderLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Issued token pair",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Refresh access/refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access/refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh token payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshTokensRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New token pair",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/switch-team": {
            "post": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Switch active team and get new tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Switch active team and get new tokens",
                "parameters": [
                    {
                        "description": "Switch team payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SwitchTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New token pair with teamId",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/debug/users/plan": {
            "post": {
                "description": "Debug: update user personal plan (free|pro)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debug"
                ],
                "summary": "Debug: update user personal plan (free|pro)",
                "parameters": [
                    {
                        "description": "Update plan payload (debug only)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/debug.UpdateUserPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan updated",
                        "schema": {
                            "$ref": "#/definitions/debug.UpdateUserPlanResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not allowed in production",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.ProviderLoginRequest": {
            "description": "Provider access token payload",
            "type": "object",
            "required": [
                "accessToken"
            ],
            "properties": {
                "accessToken": {
                    "type": "string",
                    "minLength": 1
                },
                "teamId": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshTokensRequest": {
            "description": "Refresh token payload",
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "auth.SwitchTeamRequest": {
            "description": "Switch team payload",
            "type": "object",
            "required": [
                "teamId"
            ],
            "properties": {
                "teamId": {
                    "type": "string"
                }
            }
        },
        "auth.TokenPair": {
            "description": "Token pair response",
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIs..."
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIs..."
                }
            }
        },
        "auth.TokenPairResponse": {
            "description": "Token pair wrapped in response envelope",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.TokenPair"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "debug.UpdateUserPlanRequest": {
            "description": "Update plan payload (debug only)",
            "type": "object",
            "required": [
                "plan"
            ],
            "properties": {
                "plan": {
                    "type": "string",
                    "enum": [
                        "free",
                        "pro"
                    ]
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "debug.UpdateUserPlanResponse": {
            "description": "Plan updated response",
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "debug.UpdateUserPlanResponseWrapper": {
            "description": "Update user plan response wrapped in response envelope",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/debug.UpdateUserPlanResponse"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "description": "Error response structure",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    }
}